% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{BayFDR}
\alias{BayFDR}
\title{calculate the threshold for gamma / delta for FDR control}
\usage{
BayFDR(PPI, alpha)
}
\arguments{
\item{PPI}{posterior probability of inclusion}

\item{alpha}{usually 0.05}
}
\description{
calculate the threshold for gamma / delta for FDR control
}
