% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{delta_ROC}
\alias{delta_ROC}
\title{Draw ROC curve}
\usage{
delta_ROC(deltaPPI.raw, delta.true)
}
\arguments{
\item{deltaPPI.raw}{raw values of PPI}

\item{delta.true}{XXX}
}
\description{
Draw ROC curve
}
