% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gamma_VS}
\alias{gamma_VS}
\title{Visualizition functions}
\usage{
gamma_VS(gammaPPI.raw, gamma.true, sig.level = 0.05)
}
\arguments{
\item{gammaPPI.raw}{vector of PPI}

\item{gamma.true}{XXX}

\item{sig.level}{significant level}
}
\description{
Visualizition functions
}
