% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sizefactor.estimator}
\alias{sizefactor.estimator}
\title{estimate size factor s(vector)}
\usage{
sizefactor.estimator(Ycount)
}
\arguments{
\item{Ycount}{a n-by-p count matrix Y, where n is the number of samples and p is the number of taxa(feature)}
}
\description{
estimate size factor s(vector)
}
