% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Y.filter}
\alias{Y.filter}
\title{Fitler count matrix}
\usage{
Y.filter(Ycount, zvec, min.number = 2)
}
\arguments{
\item{Ycount}{a n-by-p count matrix Y, where n is the number of samples and p is the number of taxa(feature)}

\item{zvec}{group vector}

\item{min.number}{minimal read counts}
}
\description{
Drop off the feature has less than 2 observations for each sample group
}
