% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{zinb_w_cov}
\alias{zinb_w_cov}
\title{Main function}
\usage{
zinb_w_cov(Y_mat, X_mat, z_vec, s_vec, tau_mukj = 1, S = 20000L,
  burn_rate = 0.5, tau_mu0 = 1, tau_phi = 1, tau_beta = 1,
  a_omega = 1, b_omega = 1, a_pi = 1, b_pi = 1, a_p = 1,
  b_p = 1, a_mu = 2, b_mu = 10, a_phi = 1, b_phi = 0.01,
  a_beta = 2, b_beta = 10, mu0_mean = 5, beta_lim = 5)
}
\arguments{
\item{Y_mat, }{abundance matrix}

\item{X_mat, }{covariate matrix}

\item{z_vec, }{group vector}

\item{s_vec, }{scaling factor}

\item{tau_mukj}{default = 1,}

\item{S}{default = 20000,}

\item{burn_rate}{default = 0.5,}

\item{tau_mu0}{= 1,}

\item{tau_phi}{= 1,}

\item{tau_beta}{= 1,}

\item{a_omega}{default = 0.2, uniform prior}

\item{b_omega}{defeault = 1.8, uniform prior}

\item{a_pi}{default = 1, control the sparsity of R matrix}

\item{b_pi}{default = 0.1, control the sparsity of R matrix}

\item{a_p}{default = 0.4, control the sparsity of Delta matrix}

\item{b_p}{default = 1.6, control the sparsity of Delta matrix}

\item{a_mu}{default = 2, spike and slab for mu_kj}

\item{b_mu}{default = 15, spike and slab for mu_kj}

\item{a_phi}{default = 10, dispersion parameter}

\item{b_phi}{default = 1, dispersion parameter}

\item{a_beta}{default = 2, spike and slab for beta_rj}

\item{b_beta}{default = 15, spike and slab for beta_rj}

\item{mu0_mean, }{mean}

\item{beta_lim}{= 5, limit of beta}
}
\description{
Main function
}
